refactor: `useDatePicker` composable for common reactive logic
